jill-matrix-419: README corrected; code docs corrected; refactored out common pieces for url request composing for different services.
